Generate Avro packet file
sample.rs can now emit an ingestion batch header and arbitrarily many
data share packets. It emits ingestion data for both the PHA and
facilitator servers.

I tried for a long time to get a generic interface over MemoryTransport
and FileTransport working, but I can't quite get the generic put() on
MemoryTransport to work because of issues with the lifetime of the
boxed object. I have given up for now and just made tests use temporary
dirs that get cleaned up automatically (testing was the only point of
MemoryTransport to begin with).