Move default keys to new test_utils module
No production code should be referencing these keys since this is a
security and privacy hazard. Moving them to a separate makes it slightly
easier to identify usages, but, more importantly, more easily allows us
to restrict access to them using visibility modifiers. Unfortunately
right now they are used as default keys for the sample generation so
this cannot be done yet.