Remove ConversionError
`ConversionError` was only used to wrap `TryFromIntError` in a few
specific places in the code. Since in those cases it was the only
error that could be returned, it was removed and the code in question
just returns those directly. Further up the chain they are simply
incorporated into an `anyhow::Error` as the root cause.