Use `anyhow` for most of the error propagation
`anyhow` has a single opaque error type, which can be converted from
any other type implementing the `Error` trait. More importantly, it
allows for convenient string error and adding of context to calls, and
properly keeps track of the cause chain while doing so.

A few files in the codebase weren't updated yet (mostly because of their
size) but this can be done in follow up changes.

To ease migration I added a new variant to our `prio::Error` enum, which
itself wraps `anyhow::Error` and transparently passes calls to it along.
This allows both kinds of errors to be embedded in both directions,
allowing free transit through the codebase without having to update
every function signature and error return at the same time. After the
migration is done this variant can be deleted.

My approach was to mostly get rid of the pre-existing errors we had
defined, and replace them with ad-hoc string errors (created using the
`anyhow!` macro) or `.context` calls (which are equivalent to a string
error with the original error as its cause). This is because the
existing errors mostly fell into two categories:

 - They only served to wrap a specific library error type, while adding
   a custom string message to it. The string message can be replaced by
   anyhow `.context` calls serving the same purpose, and the original
   error is preserved in the cause chain and could be matched against if
   desired. Therefore there is no need to use a wrapper error for these.
 - They were for issues arising from our own code, but at no point we
   attempted to do any special handling of them whatsoever, rendering
   them mostly informational or for display purposes. Since ad-hoc
   string errors work just as well for that they can be used instead.
   Furthermore, usage of these errors tended to be inconsistent since
   they had no clear definition of scope.

The only exception is the `EofError` kind. This is pattern-matched
against in various places and changes the behavior of the code. While I
mostly didn't touch it in this commit, I feel like the approaches for
dealing with it should be to:

 - Change functions returning it to instead return a `Result<Option<T>>`
   instead. This separates actual error cases from EOF, and can simplify
   the code handling and generating these results. This was not done
   here to keep the scope of this change small.
 - Maintain it as an error type, and then use `anyhow::Error::is` to
   check if it is in the cause chain of any returned errors.
 - Make it a more local error type, and return it in functions where EOF
   is expected, clearly differentiating it from unexpected errors by
   making those a separate variant which wraps `anyhow::Error`.