batch: Make PhantomData non-references
Maybe this struct should be restructured to not have the `H` and `P`
parameters, but in the meantime this avoids restricting variance and
tying them to the (unrelated) `'a` lifetime.

`PhantomData<*const T>` is used which doesn't imply any ownership of
data (unlike `T`) but also has no lifetimes associated with it (unlike
`&'a T`).