terraform: serve global manifest from Cloud CDN
Adds a manifest module, which configures Google Cloud CDN to serve
content from a Cloud Storage bucket over HTTPS with a managed TLS
certificate, and places the data share processor's global manifest in
the bucket. The demo config uses a domain donated by jrenken, which he
also configured a DNS A record for. In a real deploy, we would want the
domain and its DNS configuration to be managed by Terraform, which we
could achieve using a few more Googel Cloud products.