terraform: create sum part writer service account
As of just recently, MITRE/the portal server has assumed responsibility
for ownership of the buckets to which data share processors shall write
sum part batches. To enable them to grant write permissions to our
servers, we create a GCP service account, and advertise it from our
global manifest. We then allow all the service accounts we create for
individual data share processor instances to impersonate that account.