Reintroduce signature message
When evaluating message signatures, servers need to figure out what
public key to verify the message with. Thus we re-introduce the Avro
message for signatures, which now includes a key-identifier which can be
used to look up a trusted public key in the peer's global (ingestion
servers) or specific (data share processors) manifest.