facilitator: use PrioBatchSignature
Teaches facilitator to read PrioBatchSignature messages when reading
batches, and to emit PrioBatchSignatures when writing them. Also extends
command line interface to take an identifier for private keys, and
introduces the BatchSigningKey type to represent a key plus its
identifier. This commit intentionally does not include looking up a
peer's public key based on the key identifier in a signature message, on
the premise that this responsibility could be restricted to the yet
unwritten workflow manager: when scheduling a job for an individual
step, it would look up the appropriate public key with which to verify a
batch and then provide it and only it to the invocation of facilitator,
absolving facilitator of the need to consult any peer manifests.

This commit also changes the signature format we use from fixed length
PKCS#11 signatures to ASN.1 encoded signatures, as specified in
https://docs.google.com/document/d/1MdfM3QT63ISU70l63bwzTrxr93Z7Tv7EDjLfammzo6Q/edit#heading=h.3eppef3muhmr