Initial Go workflow-manager.
This adds a workflow manager that can look for batches in GCS, and can
talk to the Kubernetes API (currently just a demonstration of listing
pods).

This adds a Dockerfile to build the workflow-manager container, and
changes to kubernetes.tf to set up the appropriate role and role
bindings for workflow-manager.

The demonstration cron job used to run `facilitator`; in this PR I
change it to run `workflow-manager`. I tweaked the successful job count
because I felt that made it more explicit how many jobs I should see
sticking around.