facilitator: multiple packet decryption keys
Since not all packets in a batch are necessarily encrypted with the same
packet encryption key, the facilitator intake and aggregation setps need
to deal with multiple encryption keys. Further, since ingestors won't
always populate the encryption_key_id field (#73), we have to try all
available packet decryption keys until we find one that works (good
thing we picked an AEAD cipher so we can reliably detect that we're
using the wrong key to decrypt a message). A list of packet decryption
keys can now be provided to facilitator, either at the command line or
by an environment variable. The keys are base 64 encoded and separated
by commas.

This commit does not implement looking up which decryption key to use
with encryption_key_id if it is present. We have to get the case where
it's absent right anyway, so we may as well use that logic all the time
and save ourselves a little code and some test cases.