facilitator: look up key_identifier in map of keys
This commit teaches batch.rs to use the key_identifier value in a
signature message to look up a public key in a provided HashMap of keys.
The top level structs or methods on the intake or aggregation actions
also now expect HashMaps of keys. This includes the aggregation action:
previously it assumed that whatever key it was using to sign emitted sum
part messages should also be used to verify its own validation batches.
But since validation batches can be understood as a message from the
data share processor in the past to itself in the present, we need the
full map of public keys in order to be able to read validation batches.