Tweak facilitator args help.
Previously we had a lot of args that all had the same (long) help
explaining about GCS vs S3 storage or auth. It made it hard to read the
help because it wasn't clear that each stanza was identical. It also
obscured the most relevant information for each flag (the entity type,
i.e. peer, ingestor, self, or portal).

This moves that detailed help text out into a single staza at the top,
and shortens the help text for each flag.

I also renamed all the "ingestion-bucket" args to "ingestor-bucket",
because the bucket is owned by an ingestor. We should describe what we
do to that bucket as "intake," not "ingestion" to avoid confusion with
the "ingestor" term.

Renamed batch-intake to intake-batch so it's a verb like the "aggregate"
subcommand.