Collapse AWS and GCP identities.
Previously, for each bucket-related argument, we had two flags: One
was a boolean controlling whether we consult $AWS_ROLE_ARN to
impersonate an IAM role, and the other was the name of a GCP service
account.

These flags both contained and identity, and were mutually exclusive, so
I've collapsed them into a single flag per bucket: `--foo-identity`.

I've also abstracted out the notion of an "Entity": Ingestor, Peer, Self
(or "Own"), and Portal. This makes the calls to the argument adding
methods simpler and more expressive. There's also a new enum for the
role of a bucket in the operation: input or output.