terraform: Add IAM policy for bucket role
For each data share processor, we create an AWS IAM role which is used
by that DSP to access various S3 buckets. We also create a bucket to
which peer DSPs will write their validation batches. We aim to grant the
entire peer AWS account access to that bucket, which we achieve with
Principal { AWS { aws-account-id } } in the _bucket_ policy. It turns
out we also need a _role_ policy permitting use of the S3 API.

Issue #140 tracks restricting this policy further.