terraform: create fake server resources
In test env setups, we need someplace to write sum parts, and we need
fake manifests for the ingestion and portal servers. Based on the value
of the test_peer_environment variable, this commit will create those
resources during apply-bootstrap:
 - Output GCS bucket for sum parts
 - Portal server global manifest vended from this env's manifest bucket
 - Ingestor global manifests vended from this env's manifest bucket

This requires that the ingestors and portal_server_manifest_base_url
variables in the env's .tfvars file point to the URL or bucket where
these resources will be created. The ingestor global manifests will
advertise a fixed public key corresponding to the batch signing key that
the sample_maker jobs will later be configured to use, once a
non-bootstrap apply is run.