workflow-manager: make build more self-contained.
Taking some ideas from #168, this builds off a Go image instead of
relying on the person building to have Go installed. It uses go mod
download to build a layer with dependencies cached.

This still relies on a build.sh run on the host, which gathers
information about the git branch. To streamline things, it combines
all the build info into a single variable.