facilitator: use string for GCP service account ID
Google is a big data company, so the numeric unique IDs for GCP service
accounts don't fit into u64. While that value is an integer, it is
actually opaque and meaningless to the facilitator, serving only to be
plugged into an AWS role assumption IAM policy. Accordingly we now
handle it as an Option<String>.