workflow-manager: make build more self-contained. (#173)
Taking some ideas from #168, this builds off a Go image instead of
relying on the person building to have Go installed. It uses go mod
download to build a layer with dependencies cached.

This still relies on a build.sh run on the host, which gathers
information about the git branch. To streamline things, it combines
all the build info into a single variable.

Sets the build timestamp to local time on the build machine. This
is easier to reference when building on a dev workstation; for release
builds this will be whatever GH Actions sets, which I'm pretty sure is UTC.

Attach a label with the build info to the image. This lets us check
the metadata for an image in a registry to confirm when it was built
and what branch.