Advertise ingestion writer role ARN from manifest
In order for ingestors in GCP to be able to write to ingestion buckets,
they must know what AWS IAM role to assume. We were already creating the
necessary role and configuring ingestion buckets to allow it to write,
but we hadn't exposed the ARN in the specific manifest, which this
commit fixes.