facilitator: support GCP SA key files
To enable our disginguished colleagues at NCI to write to Google Cloud
Storage buckets from outside of GCP, we add support for using a GCP
service account key file to obtain OAuth tokens for the default service
account. Previously, we only supported fetching an OAuth token from the
GKE metadata service, but now we can be provided the path to a service
account key file (as one might obtain through the Google Cloud console),
and will use the service account email and private key therein to
construct a signed JWT and then send a request to an OAuth API endpoint
to get an OAuth token suitable for authenticating in turn to the Google
Cloud Storage service.

This commit also fixes a bug where the logic for determining if an OAuth
token has expired was exactly backwards.