facilitator: Add AWS ChainProvider for WebIDP.
Rusoto offers a "ChainProvider", which will try each of the various auth
methods available on AWS in order, and return as soon as one works. The
ChainProvider offered by Rusoto omits WebIdentityProvider, which we
need.

This change copies the implementation of ChainProvider, and adds
WebIdentityProvider to the list of providers tried.