workflow-manager: include tag in build info.
At build time, we embed the branch and git commit into workflow-manager,
so it can print that when it starts up. However, the `git symbolic-ref`
invocation doesn't work when checked out to a tag rather than a branch.
That meant that our tagged releases wouldn't have an informative build
ID. This adds handling for tags.