facilitator: same BUILD_INFO as workflow-manager. (#215)
For workflow-manager, we've done some work to determine the branch name OR
the release tag and inject them into the build binary. This ports that
change over to facilitator, getting rid of our vergen dependency.

This introduces a build.sh script for facilitator that passes a
BUILD_INFO arg into the Docker build. The Dockerfile passes that along
as an environment variable, which is used by rustc at compile time.