terraform: Subscribe to a GKE release channel
This enables the cluster to be auto-updated to new versions at a regular
cadence. See [1] for more information on GKE release channels.

The documentation around this is unclear, but it seems that if a release
channel is not specified, the cluster doesn't default to any of the
regular channels, and instead, auto-updates only happen on an
as-required basis without a clear schedule.[2]

[1]: https://cloud.google.com/kubernetes-engine/docs/concepts/release-channels
[2]: https://cloud.google.com/kubernetes-engine/docs/release-notes#no-channel