Pin gamlin-test/narnia env to 0.4.8 containers
We got through Narnia tests with the 0.4.8 release of workflow-manager
and facilitator. This commit pins the gamlin-test environment to that
version (instead of pulling latest) so that we can continue to release
newer bits from main without breaking the test env, which we will keep
around for a bit longer in case our GAMLIN partners want to continue to
push data through it.