terraform: Use non-beta google provider by default
Almost all of our GCP resources explicitly overrided the provider to use
the google-beta provider. This provider variant was needed to access
some of features of GCP that the provider considers non-GA.

However, there's no requirement that every resource use either only
google or google-beta. The two provider variants interoperate fine and
google-beta can be opted-in only on an as-needed basis. The only thing
that was missing to allow this was a configuration block for the
non-beta variant of the provider, to specify default values.

This removes the need to explicitly declare the provider in each
individual resource and removes the possibility of accidentaly
forgetting to do so (which existed as a bug in the current version of
the code, and which tends to lead to confusing error messages or strange
behavior due to the default parameters such as project not being set).