terraform: configure Prometheus provider
Make sure the Prometheus provider uses credentials for the Kubernetes
cluster managed by Terraform, and not ambient credentials from ~/.kube,
which may not be for the right cluster. Additionally, we now set the
repository from which to pull the Prometheus chart so that operators
don't need to `helm add stable` outside of Terraform before deploying.