deal with non-singleton ingestors
Google will have one ingestion server per locality, using distinct batch
signing keys and GCP service accounts. We now will check for the
ingestor global manifest, and if it cannot be found, will fall back to a
specific manifest at a path "<instance-name>-manifest.json", i.e.,
"us-co-g-enpa-manifest.json". If there is a global manifest but it is
not a valid manifest, we still error out.

resolves #107