Add Prometheus metrics to facilitator (#252)
This adds some very basic metrics - number of intake-batch and aggregate
jobs started. The main intent is to set up the scaffolding for pushing
metrics from facilitator. This adds a variable for the address of the
pushgateway. If not address is provided (the default), no metrics are
pushed. A similar scheme should be adopted for workflow-manager (#251).

This incidentally factors out some of the arguments to BatchIntaker and
BatchAggregator to reduce nesting a bit.