facilitator: optionally drop nth packet in samples
This adds a parameter to the generate_ingestion_sample function that
makes it drop the nth packet from ingestion batches generated, enabling
testing of cases where the PHA and facilitator ingestion batches are
inconsistent. We also now record which packets were dropped from either
side as well as how many packets made it into both batches, to enable
tests to figure out what the expected sums and contribution counts are.