workflow-manager: enqueue tasks in PubSub
The model in which `workflow-manager` scheduled a Kubernetes job for each intake
or aggregation job scaled poorly. This commit begins to implement a producer/
consumer model in which `workflow-manager` schedules tasks by writing messages
to Google PubSub or to AWS SNS, depending on what parameters are
provided.

Once it either schedules a task or finds a k8s job indicating that an older
`workflow-manager` had already handled a task, `workflow-manager` now writes out
a marker object to the own validation bucket, to avoid scheduling duplicate
tasks.

`workflow-manager` will not delete completed batch intake jobs it
encounters, on the assumption that such deletions will be a one-time
task as deployments move from older versions to the pub-sub model and so
it's not worth carrying around code to support it long-term.

`workflow-manager` can now be provided a kubeconfig file so that it can be run
outside of a Kubernetes cluster; see `workflow-manager/README.md` for details.

`workflow-manager` now has "dry run" mode in which it will avoid any operations
with side effects (e.g., Kubernetes job deletion or writing to buckets) and
instead log what it would have done. See `workflow-manager/README.md` for more.

Finally, this commit refactors `main()` so that the core task scheduling logic
can at last be tested.