workflow-manager: additional metrics
Fixes a bug where metrics were being pushed to the pushgateway at the
beginning of the workflow-manager instead of the end, such that no
metrics were pushed before being gathered locally.

Additionally adds new gauges for:
 - number of tasks scheduled
 - number of tasks detected but skipped due to age, task marker or
   existing k8s job
 - time of last successful and failed runs
 - how long it took to run workflow-manager

All of these are gauges, in keeping with the advice in Prometheus
documentation (https://prometheus.io/docs/practices/instrumentation).
Because this is a batch job, we can't meaningfully use anything but
gauges, as other collector types would get reset with each run of
workflow-manager.

Existing metrics are renamed to prefix them with `workflow_manager_`,
and all metrics are grouped under the `locality` label, which is
populated with the Kubernetes namespace name. `instance` is not set,
which will cause pushgateway to set it to `""`. This is the desired
behavior, since there are no persistent instances of `workflow-manager`.