workflow-manager: ignore task markers
The new version of workflow-manager introduced in #292 writes out task
marker objects to the own validation bucket to track which intake and
aggregation tasks have been scheduled. This commit teaches the old
version of workflow-manager to gracefully ignore any such files, to give
us something safe to roll back to in case deploying #292 to production
fails in such a way that leaves task markers behind.