terraform: adjust workflow-manager resource limits
Raises memory and CPU limits for the workflow-manager so that long lists
of Kubernetes jobs and cloud storage bucket listings can fit safely in
memory. The CPU limits are raised just to let the workflow-manager go a
little faster, though realistically it's going to spend most of its time
waiting on API calls.