facilitator: extend task deadlines
Adds methods to the `TaskQueue` trait that allow extending task
deadlines (aka ack deadline in GCP PubSub or visibility timeout in AWS
SQS) when tassk are running long. Also adds callback parameters to
`BatchIntaker::generate_validation_share` and
`BatchAggregator::generate_sum_part` which fire periodically to give a
caller an opportunity to evaluate how much time has elapsed since the
start of the task and extend deadlines as necessary. See #308 for more
context on why this is needed.

Resolves #308