workflow-manager: relax timeout on GCS object list
We observed workflow-manager failing in the `us-md-apple` instance
because we were hitting timeouts enumerating the contents of GCS
buckets. This is because the `list` API on buckets is paginated
(https://cloud.google.com/storage/docs/json_api/v1/buckets/list), which
the Go SDK handles by returning an `ObjectIterator` from
`BucketHandle.Objects()`. However you can only provide a
`context.Context` to the initial call to `Objects()`, and each
subsequent call to `ObjectIterator.next()` is subject to the same
initial timeout. There's no way to reset the timeout on a
`context.Context` that I can find, so I believe the thing to do is to
raise the timeout for this particular operation, which we understand
actually does take longer than 30 seconds, to 10 minutes. This should be
more than enough time to list the contents of buckets, while still
allowing us to eventually give up on stuck `workflow-manager` jobs.

Additionally, while looking into correct handling of timeouts in GCS, I
found mention in Google documentation that timeouts should not be
applied when calling the various `NewClient()` functions from the GCP
SDK. Accordingly those calls are amended to use `context.Background()`.