workflow-manager: label metrics by ingestor
`workflow-manager` metrics were grouped by namespace (hence, locality),
but metrics emitted by either ingestor's `workflow-manager` were going
to the same time series, leaving us unable to distinguish between the
health of Apple vs. g-enpa task schedulers. This commit adds a new
grouping to metrics pushed by `workflow-manager` for ingestor name, and
amends Terraform to pass that ingestor name into the `workflow-manager`.

This also fixes a problem with the `prometheus-server` deployment: by
default, the Prometheus helm chart creates a Kubernetes deployment for
`prometheus-server` with an update strategy of `RollingUpdate`. This
doesn't work with a persistent volume because it cannot be shared across
two revisions during an update. We opt into the `Recreate` strategy to
work around this.

Resolves #317