terraform: configurable intake-max-age
The `intake-max-age` parameter to `workflow-manager` had been hard-coded
to 24h. This is longer than makes sense given our aggregation window of
8 hours in production, and in any case this should be configurable,
since being able to tune that parameter has already helped us out of a
couple of different incidents. This commit adds a top level Terraform
variable for `intake-max-age`, and sets the default to the more
reasonable value of six hours.

Resolves #331