facilitator: don't abort integration for bad batch
We are losing whole aggregations because one peer validation batch out
of ~900 happens to have an invalid signature or an incorrect packet file
digest in its header. We should still be able to aggregate the other 899
batches in such a case. We now continue with the aggregation if an
individual call to `BatchAggregator::aggregate_share()` fails.

We also now gather metrics on how many invalid own and peer validation
batches are encountered during an aggregation. This will enable us to
get a better idea of how often the bug that caused invalid validation
batches occurs.

Resolves #343
Resolves #345