facilitator: use long timeouts on GCS uploads
This commit increases the timeout used for GCS uploads to 120 seconds
from 10 seconds. 10 seconds appeared to suffice for uploading to GCS
from within GCP, but log analysis shows that NCI intake jobs were
sometimes failing because of this timeout, in a way that your humble
commit author believes was ultimately causing the aggregation failures
we have been seeing.