facilitator: structured error handling
This change is the start of more rigorous error handling in
`facilitator`, making better use of Rust enums to classify errors. We
introduce several new variants to the `Error` enum to specifically
capture errors around parsing or validating Avro batches. This enables
`BatchAggregator` to report when a failure is due to invalid data, in
which case we do not want to retry. We also can now tell when a failure
was due to bad data from a peer (report this but don't alert) or from
ourselves (do alert).

Over time, we should eliminate most if not all uses of `anyhow!` and the
`.context()` method from `Anyhow::Context` in favour of introducing
specific variants on `crate::Error`, and have all module interfaces
return `Result<T, crate::Error>` rather than the implicit `Result<T,
anyhow::Error>`.

See https://nick.groenen.me/posts/rust-error-handling/ for stimulating
discussion of library vs. application error handling in Rust and that
intersects with usage of `thiserror` vs. `anyhow`.