workflow-manager: set success metric
After the 2021/2/18 production deploy, we encountered alerts because the
`workflow_manager_last_success_seconds` metric was not being set in
`ta-ta-apple`, `ta-ta-g-enpa` or `us-va-g-enpa`. This is because no data
comes into those instances, and so no aggregation IDs are discovered in
their ingestion buckets, and due to a bug introduced by #370, the metric
was being set only if there were tasks to schedule. We had the same
problem with the `workflow_manager_runtime_seconds` metric, which would
get set too many times during a run of `workflow-manager` (though the
last value that would get set was correct).

We now also configure the staging environment to _not_ create a
`sample-maker` cronjob in one of the three localities it creates, so
that going forward, our staging environment can reproduce the issue we
saw in production.