terraform: bump task_queue_growth alert to 3h
We have observed that it can sometimes take just over two hours to get
through a set of aggregation tasks, especially if one of them is failing
in a way that causes five retries (imagine a case where an aggregation
task will fail on the 716th batch, but it takes 25 minutes of work to
get to the 716th, and you retry five times before giving up). Further,
now that we are turning on California, which has quite a lot of users, I
suspect the aggregation jobs may legitimately take longer than 2h. To
reduce the chance of spurious, unactionable alerts, this commit changes
the `task_queue_growth` alert to fire after a queue is nonempty for
three hours.