facilitator: refactor AWS authentication
`facilitator` includes clients for AWS SQS and S3. SQS is exclusively
used by deployments on AWS, while the S3 client is also used by
deployments on GCP to write to peer-owned buckets in S3. This means we
have three different expected means of authenticating to AWS:

 - deployments in Google Kubernetes Engine use a Rusoto
   `WebIdentityProvider` to obtain Oauth tokens from the GKE metadata
   service;
 - deployments in AWS Elastic Kubernetes Service and/or Fargate use a
   Rusoto WebIdentityProvider configured with environment variables
   injected by EKS (`WebIdentityProvider::from_k8s_env()`);
 - deployments in other AWS contexts (possibly including EKS without
   Fargate, I'm not sure) use Rusoto's `DefaultCredentialsProvider` to
   obtain credentials from either AWS environment variables, instance
   metadata or `~/.aws/credentials`.

Previously, we used a copied-and-slightly-modified version of Rusoto's
`DefaultCredentialsProvider` that chained a `WebIdentityProvider` along
with other credentials providers. This is bad because in NCI's Fargate
deployment, only the Kubernetes environment WebIdentityProvider could
possibly work, and trying to fall back to other providers caused us to
discard the error from the WebIdP and not retry.

This commit makes the default behavior on AWS to exclusively try to use
a `WebIdentityProvider::from_k8s_env()` (to maintain backward
compatibility with NCI's deployment), and allows clients to explicitly
opt into using a `DefaultCredentialsProvider` (to support the incoming
MITRE deployment for ENPA international).