terraform: provision 8 us-ut-apple intake workers
We got paged on 2021/3/10 because the `us-ut/intake-batch-apple`
deployment wasn't keeping up with the incoming tasks because of a spike
in traffic. I resolved that by manually scaling the deployment up to 8
workers, from 5. We observed another similar spike in incoming ingestion
batches around 4 AM PST 2021/3/11, so I want to keep the 8 workers until
we get around to autoscaling the workers (#484).