facilitator: multi-threaded aggregation
We observed that in NCI's production deployment, us-ca-apple
individual aggregation tasks are often taking more than eight hours to
run, which means that even adding more aggregation workers won't speed
things up. We schedule aggregations to run every eight hours, making it
impossible for NCI to keep up with incoming work.

To ameliorate this, this commit refactors the aggregation task handler
so that it can spawn multiple threads and sum over individual batches in
parallel. The threads are managed using the `WorkQueue` struct, which
implements a no-producer, multiple-consumer work queue. I say
"no-producer" because it doesn't support an "enqueue" method as you
might see in most pub/sub implementations, and instead expects the
entire list of jobs to be available at the creation of the `WorkQueue`.
In our use case, we know up front the list of jobs we want to execute in
parallel, and this allows us to greatly simplify the queue
implementation. For instance, we can use a simple `Vec` to store jobs
instead of requiring a `VecDeque` or some more complex data structure.

To allow safely sharing or sending resources across worker threads, we
rely heavily on structs implementing the `Clone` trait. Mostly this is a
simple matter of tagging existing definitions with `#[derive(Clone)]`,
but it required some extra gymnastics for `S3Client` and its associated
structs:

	- `ClientProvider` was refactored from a supertrait over
	  `std::ops::Fn` into a more conventional trait to make it
	  easier to reason about how to `Clone` it.
	- Because Rusoto's `MockRequestDispatcher` & Co. are not
	  `Clone` and have private fields, we had to introduce some
	  proxy objects that can be cloned.

Additionally, `AwsSqsTaskQueue` previously had a
`tokio::runtime::Runtime` member. Since `Runtime` or
`tokio::runtime::Handle` can't be cloned with the semantics that we
want, we instead instantiate a fresh `Runtime` for every API request.
[See the discussion thread in the pull request for more
details](https://github.com/abetterinternet/prio-server/pull/486#discussion_r595574887).

To enable multithreaded aggregation, pass `--thread-count=<int>` to
`facilitator aggregate-worker`, or set the `THREAD_COUNT` environment
variable to an integer value (e.g., in a Kubernetes ConfigMap).

This commit also introduces the notion of a Watchdog.

We have long supported the notion of a callback on the batch intake and
aggregate tasks, which were used to implement extending deadlines on GCP
PubSub subscriptions and AWS SQS queues. Modeling this as a closure made
it hard to reason about the ownership, thread safety, cloneability, etc.
of the values closed over by the callbacks, so I refactored it into a
`Watchdog` trait, with implementations:

 - `NoOpWatchdog`, used in `intake_batch_subcommand` and
   `aggregate_subcommand`
 - `TaskQueueWatchdog`, used in `intake_batch_worker` and
   `aggregate_worker`
 - `IncrementingWatchdog`, used in `integration_tests.rs`.

`Watchdog` is arguably a misnomer: a real watchdog would detect that too
long had passed since the last heartbeat and then do something like
alert or kill the process, but all the `Watchdog` implementations added in
this commit only get to run when `send_heartbeat` is called.

`TaskQueueWatchdog` could be extended in the future so that it runs in its
own thread and does that sort of thing, perhaps as part of addressing #293.

Resolves #319